feat: BoxOffice DTO 코드 완성
